Revert "LPS-59737 Standardize package for Control Menu Entries"
This reverts commit 5314dadf1e94cf5a3f406e000d5f440a7c6c26b8.
