+++ modules/apps/control-menu/control-menu-web/src/main/java/com/liferay/control/menu/web/InformationMessagesControlMenuEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.control.menu.web.control.menu.entry;
+package com.liferay.control.menu.web;
 
 import com.liferay.control.menu.BaseJSPControlMenuEntry;
 import com.liferay.control.menu.ControlMenuEntry;
