+++ modules/apps/control-menu/control-menu-web/src/main/java/com/liferay/control/menu/web/ManageLayoutControlMenuEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.control.menu.web.control.menu.entry;
+package com.liferay.control.menu.web;
 
 import com.liferay.control.menu.BaseControlMenuEntry;
 import com.liferay.control.menu.ControlMenuEntry;
