+++ modules/apps/control-menu/control-menu-web/src/main/resources/META-INF/resources/view.jsp
@@ -65,9 +65,9 @@ if (layout != null) {
 
 							<li>
 								<liferay-ui:icon
-									iconCssClass='<%= controlMenuEntry.getIconCssClass(request) + " control-menu-icon icon-monospaced" %>'
+									iconCssClass='<%= controlMenuEntry.getIconCssClass(request) + " icon-monospaced" %>'
 									label="<%= false %>"
-									linkCssClass="<%= controlMenuEntry.getLinkCssClass(request) %>"
+									linkCssClass='<%= "control-menu-icon " + controlMenuEntry.getLinkCssClass(request) %>'
 									message="<%= controlMenuEntry.getLabel(locale) %>"
 									url="<%= controlMenuEntry.getURL(request) %>"
 									useDialog="<%= controlMenuEntry.isUseDialog() %>"
